Add type hints
This was more complicated than I anticipated because of
Python 2 support.
Values extracted from JSON or TOML are Unicode.  Values derived
from JSON or TOML are also Unicode.  PEP 517 inputs and outputs
are all Unicode.  Input paths are required to be in Unicode
for consistency with PEP 517.
The subprocess runner can take either bytes or Unicode as an argument
to `env`, and bytes in Python 2 and Unicode in Python 3
as an argument to `cmd`.

Closes #90.